Refactor same-month calculus workflow
